gcp_upload_function/requirements.txt
google-cloud-storage
requests
